requirements.txt
requests
progressbar2
bs4